requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
huggingface_hub==0.27.1
pytest==8.3.3
respx==0.21.1
